Moved batch related stuff to own package
